Rename FixedTimestampIntegrator to FixedTimestepIntegrator
