+++ src/main.tsx
@@ -1,4 +1,4 @@
-import { FixedTimestepIntegrator } from './engine/FixedTimestampIntegrator';
+import { FixedTimestepIntegrator } from './engine/FixedTimestepIntegrator';
 import { Grid } from './engine/Grid';
 import { Vec2D } from './engine/Vec2D';
 import { Viewport } from './engine/Viewport';
